v2 document picker complete
